Zero functional changes, just renumbering (but not reordering) migrations to leave gaps for continued refinement prior to our first "release".
