+++ migration/src/m0000020_create_cvss4_enums.rs
