+++ migration/src/m0000050_create_vulnerability_description.rs
@@ -1,6 +1,8 @@
-use crate::m0000030_create_advisory::Advisory;
-use crate::m0000042_create_package_version::PackageVersion;
+use crate::m0000040_create_vulnerability::Vulnerability;
+use crate::m0000090_create_advisory_vulnerability::AdvisoryVulnerability;
 use sea_orm_migration::prelude::*;
+
+use crate::Now;
 
 #[derive(DeriveMigrationName)]
 pub struct Migration;
@@ -12,34 +14,40 @@ impl MigrationTrait for Migration {
         manager
             .create_table(
                 Table::create()
-                    .table(NotAffectedPackageVersion::Table)
+                    .table(VulnerabilityDescription::Table)
                     .if_not_exists()
                     .col(
-                        ColumnDef::new(NotAffectedPackageVersion::Id)
+                        ColumnDef::new(VulnerabilityDescription::Id)
                             .integer()
                             .not_null()
                             .auto_increment()
                             .primary_key(),
                     )
                     .col(
-                        ColumnDef::new(NotAffectedPackageVersion::AdvisoryId)
+                        ColumnDef::new(AdvisoryVulnerability::VulnerabilityId)
                             .integer()
                             .not_null(),
                     )
                     .foreign_key(
                         ForeignKey::create()
-                            .from_col(NotAffectedPackageVersion::AdvisoryId)
-                            .to(Advisory::Table, Advisory::Id),
+                            .from_col(VulnerabilityDescription::VulnerabilityId)
+                            .to(Vulnerability::Table, Vulnerability::Id)
+                            .on_delete(ForeignKeyAction::Cascade),
                     )
                     .col(
-                        ColumnDef::new(NotAffectedPackageVersion::PackageVersionId)
-                            .integer()
+                        ColumnDef::new(VulnerabilityDescription::Timestamp)
+                            .timestamp_with_time_zone()
+                            .default(Func::cust(Now)),
+                    )
+                    .col(
+                        ColumnDef::new(VulnerabilityDescription::Lang)
+                            .string()
                             .not_null(),
                     )
-                    .foreign_key(
-                        ForeignKey::create()
-                            .from_col(NotAffectedPackageVersion::PackageVersionId)
-                            .to(PackageVersion::Table, PackageVersion::Id),
+                    .col(
+                        ColumnDef::new(VulnerabilityDescription::Description)
+                            .string()
+                            .not_null(),
                     )
                     .to_owned(),
             )
@@ -50,7 +58,7 @@ impl MigrationTrait for Migration {
         manager
             .drop_table(
                 Table::drop()
-                    .table(NotAffectedPackageVersion::Table)
+                    .table(VulnerabilityDescription::Table)
                     .to_owned(),
             )
             .await
@@ -58,11 +66,12 @@ impl MigrationTrait for Migration {
 }
 
 #[derive(DeriveIden)]
-pub enum NotAffectedPackageVersion {
+pub enum VulnerabilityDescription {
     Table,
     Id,
-    //Timestamp,
+    Timestamp,
     // --
-    AdvisoryId,
-    PackageVersionId,
+    VulnerabilityId,
+    Lang,
+    Description,
 }
