+++ migration/src/m0000060_create_advisory.rs
