+++ migration/src/m0000090_create_advisory_vulnerability.rs
@@ -1,8 +1,6 @@
-use crate::m0000011_create_vulnerability::Vulnerability;
-use crate::m0000032_create_advisory_vulnerability::AdvisoryVulnerability;
+use crate::m0000040_create_vulnerability::Vulnerability;
+use crate::m0000060_create_advisory::Advisory;
 use sea_orm_migration::prelude::*;
-
-use crate::Now;
 
 #[derive(DeriveMigrationName)]
 pub struct Migration;
@@ -14,14 +12,18 @@ impl MigrationTrait for Migration {
         manager
             .create_table(
                 Table::create()
-                    .table(VulnerabilityDescription::Table)
+                    .table(AdvisoryVulnerability::Table)
                     .if_not_exists()
                     .col(
-                        ColumnDef::new(VulnerabilityDescription::Id)
+                        ColumnDef::new(AdvisoryVulnerability::AdvisoryId)
                             .integer()
-                            .not_null()
-                            .auto_increment()
-                            .primary_key(),
+                            .not_null(),
+                    )
+                    .foreign_key(
+                        ForeignKey::create()
+                            .from_col(AdvisoryVulnerability::AdvisoryId)
+                            .to(Advisory::Table, Advisory::Id)
+                            .on_delete(ForeignKeyAction::Cascade),
                     )
                     .col(
                         ColumnDef::new(AdvisoryVulnerability::VulnerabilityId)
@@ -30,24 +32,14 @@ impl MigrationTrait for Migration {
                     )
                     .foreign_key(
                         ForeignKey::create()
-                            .from_col(VulnerabilityDescription::VulnerabilityId)
+                            .from_col(AdvisoryVulnerability::VulnerabilityId)
                             .to(Vulnerability::Table, Vulnerability::Id)
                             .on_delete(ForeignKeyAction::Cascade),
                     )
-                    .col(
-                        ColumnDef::new(VulnerabilityDescription::Timestamp)
-                            .timestamp_with_time_zone()
-                            .default(Func::cust(Now)),
-                    )
-                    .col(
-                        ColumnDef::new(VulnerabilityDescription::Lang)
-                            .string()
-                            .not_null(),
-                    )
-                    .col(
-                        ColumnDef::new(VulnerabilityDescription::Description)
-                            .string()
-                            .not_null(),
+                    .primary_key(
+                        Index::create()
+                            .col(AdvisoryVulnerability::AdvisoryId)
+                            .col(AdvisoryVulnerability::VulnerabilityId),
                     )
                     .to_owned(),
             )
@@ -58,7 +50,8 @@ impl MigrationTrait for Migration {
         manager
             .drop_table(
                 Table::drop()
-                    .table(VulnerabilityDescription::Table)
+                    .table(AdvisoryVulnerability::Table)
+                    .if_exists()
                     .to_owned(),
             )
             .await
@@ -66,12 +59,8 @@ impl MigrationTrait for Migration {
 }
 
 #[derive(DeriveIden)]
-pub enum VulnerabilityDescription {
+pub enum AdvisoryVulnerability {
     Table,
-    Id,
-    Timestamp,
-    // --
+    AdvisoryId,
     VulnerabilityId,
-    Lang,
-    Description,
 }
