+++ migration/src/m0000120_create_package_version.rs
@@ -1,4 +1,4 @@
-use crate::m0000042_create_package_version::PackageVersion;
+use crate::m0000100_create_package::Package;
 use sea_orm_migration::prelude::*;
 
 use crate::Now;
@@ -13,31 +13,32 @@ impl MigrationTrait for Migration {
         manager
             .create_table(
                 Table::create()
-                    .table(QualifiedPackage::Table)
+                    .table(PackageVersion::Table)
                     .if_not_exists()
                     .col(
-                        ColumnDef::new(QualifiedPackage::Id)
+                        ColumnDef::new(PackageVersion::Id)
                             .integer()
                             .not_null()
                             .auto_increment()
                             .primary_key(),
                     )
                     .col(
-                        ColumnDef::new(QualifiedPackage::Timestamp)
+                        ColumnDef::new(PackageVersion::Timestamp)
                             .timestamp_with_time_zone()
                             .default(Func::cust(Now)),
                     )
                     .col(
-                        ColumnDef::new(QualifiedPackage::PackageVersionId)
+                        ColumnDef::new(PackageVersion::PackageId)
                             .integer()
                             .not_null(),
                     )
                     .foreign_key(
                         ForeignKey::create()
-                            .from_col(QualifiedPackage::PackageVersionId)
-                            .to(PackageVersion::Table, PackageVersion::Id)
+                            .from_col(PackageVersion::PackageId)
+                            .to(Package::Table, Package::Id)
                             .on_delete(ForeignKeyAction::Cascade),
                     )
+                    .col(ColumnDef::new(PackageVersion::Version).string().not_null())
                     .to_owned(),
             )
             .await
@@ -45,16 +46,17 @@ impl MigrationTrait for Migration {
 
     async fn down(&self, manager: &SchemaManager) -> Result<(), DbErr> {
         manager
-            .drop_table(Table::drop().table(QualifiedPackage::Table).to_owned())
+            .drop_table(Table::drop().table(PackageVersion::Table).to_owned())
             .await
     }
 }
 
 #[derive(DeriveIden)]
-pub enum QualifiedPackage {
+pub enum PackageVersion {
     Table,
     Id,
     Timestamp,
     // --
-    PackageVersionId,
+    PackageId,
+    Version,
 }
