+++ migration/src/m0000200_sbom_describes_package.rs
@@ -1,5 +1,5 @@
-use crate::m0000010_create_sbom::Sbom;
-use crate::m0000044_create_qualified_package::QualifiedPackage;
+use crate::m0000030_create_sbom::Sbom;
+use crate::m0000130_create_qualified_package::QualifiedPackage;
 use sea_orm_migration::prelude::*;
 
 #[derive(DeriveMigrationName)]
